+++ nengo_loihi/simulator.py
@@ -10,7 +10,6 @@ import nengo.utils.numpy as npext
 import numpy as np
 
 from nengo_loihi.builder import Model
-from nengo_loihi.builder.nengo_dl import HAS_DL, install_dl_builders
 from nengo_loihi.compat import NengoSimulationData
 from nengo_loihi.emulator import EmulatorInterface
 from nengo_loihi.hardware import HardwareInterface, HAS_NXSDK
@@ -114,9 +113,6 @@ class Simulator:
 
         if progress_bar:
             warnings.warn("nengo-loihi does not support progress bars")
-
-        if HAS_DL:
-            install_dl_builders()
 
         if model is None:
             self.model = Model(dt=float(dt), label="%s, dt=%f" % (network, dt))